fix(9585): scheduledtask time selection made into sections component
