+++ containers/Cloudenv/views/cloudaccount/dialogs/ScheduledtaskCreate.vue
@@ -14,59 +14,7 @@
         <a-form-item :label="$t('cloudenv.text_360')">
           {{ $t('cloudenv.sync_account') }}
         </a-form-item>
-        <a-form-item :label="$t('cloudenv.text_433')">
-          <a-radio-group v-decorator="decorators.cycle_type">
-            <a-radio-button v-for="(v, k) in $t('cloudenvScheduledtaskGroupCycleType')" :key="k" :value="k">{{v}}</a-radio-button>
-          </a-radio-group>
-        </a-form-item>
-        <!-- 周期策略 -->
-        <a-form-item :label="$t('cloudenv.frequency')" v-if="form.fc.getFieldValue('cycleTimer.cycle_type') === 'cycle'">
-          <div class="d-flex align-items-center">
-            {{$t('cloudenv.each')}}
-            <a-input-number class="ml-2" :min="1" v-decorator="decorators.cycle_num" />
-            <a-select class="mr-2" v-decorator="decorators.cycle_type2" style="width: 100px">
-              <a-select-option v-for="(item,index) in cycleTypeOpts" :key="index" :value="item.key">{{item.label}}</a-select-option>
-            </a-select>
-            {{$t('cloudenv.cycle_once')}}
-          </div>
-        </a-form-item>
-        <!-- 周 -->
-        <a-form-item :label="$t('cloudenv.text_434')" v-if="form.fc.getFieldValue('cycleTimer.cycle_type') === 'week'">
-          <a-select v-decorator="decorators.weekDays" mode="multiple">
-            <a-select-option v-for="(v, k) in $t('flexGroupSubCycleTypeWeek')" :key="k" :value="parseInt(k)">{{v}}</a-select-option>
-          </a-select>
-        </a-form-item>
-        <!-- 月 -->
-        <a-form-item :label="$t('cloudenv.text_435')" v-if="form.fc.getFieldValue('cycleTimer.cycle_type') === 'month'">
-          <a-select v-decorator="decorators.monthDays" mode="multiple">
-            <a-select-option v-for="i in 31" :key="i" :value="parseInt(i)">{{$t('cloudenv.text_436', [i])}}</a-select-option>
-          </a-select>
-        </a-form-item>
-        <!-- 单次策略 -->
-        <template v-if="form.fc.getFieldValue('cycleTimer.cycle_type') === 'one'">
-          <a-form-item :label="$t('cloudenv.text_437')">
-            <a-date-picker
-              :showTime="{
-                format: 'HH:mm',
-              }"
-              :disabledDate="disabledDate"
-              :disabledTime="disabledDateTime"
-              v-decorator="decorators.execTime"
-              format="YYYY-MM-DD HH:mm" />
-          </a-form-item>
-        </template>
-        <!-- 非单次策略 -->
-        <template v-if="form.fc.getFieldValue('cycleTimer.cycle_type') !== 'one'">
-          <a-form-item :label="$t('cloudenv.text_437')" v-if="form.fc.getFieldValue('cycleTimer.cycle_type') !== 'cycle'">
-            <a-time-picker v-decorator="decorators.hourMinute" format="HH:mm" />
-          </a-form-item>
-          <a-form-item :label="$t('cloudenv.text_438')" :help="$t('cloudenv.range_picker_help')">
-            <a-range-picker
-              v-decorator="decorators.startEndTime"
-              :disabledDate="disabledDate"
-              format="YYYY-MM-DD" />
-          </a-form-item>
-        </template>
+        <scheduled-task-time ref="taskTime" :form="form" />
       </a-form>
     </div>
     <div slot="footer">
@@ -77,6 +25,7 @@
 </template>
 
 <script>
+import ScheduledTaskTime from '@Cloudenv/sections/ScheduledTaskTime'
 import {
   getStatusTableColumn,
   getEnabledTableColumn,
@@ -89,6 +38,7 @@ export default {
   name: 'ScheduledtaskCreateDialog',
   components: {
     NameRepeated,
+    ScheduledTaskTime,
   },
   mixins: [DialogMixin, WindowsMixin],
   data () {
@@ -100,77 +50,6 @@ export default {
           {
             rules: [
               { required: true, message: `${this.$t('common.placeholder')}${this.$t('common.name')}` },
-            ],
-          },
-        ],
-        cycle_type: [
-          'cycleTimer.cycle_type',
-          {
-            initialValue: 'day',
-            rules: [
-              { required: true, message: this.$t('cloudenv.text_443') },
-            ],
-          },
-        ],
-        weekDays: [
-          'cycleTimer.week_days',
-          {
-            initialValue: [1, 2, 3, 4, 5],
-            rules: [
-              { required: true, message: this.$t('cloudenv.text_444') },
-            ],
-          },
-        ],
-        monthDays: [
-          'cycleTimer.month_days',
-          {
-            initialValue: [1],
-            rules: [
-              { required: true, message: this.$t('cloudenv.text_445') },
-            ],
-          },
-        ],
-        cycle_num: [
-          'cycleTimer.cycle_num',
-          {
-            initialValue: 1,
-            rules: [
-              { required: true, message: this.$t('cloudenv.input_cycle_num') },
-            ],
-          },
-        ],
-        cycle_type2: [
-          'cycleTimer.cycle_type2',
-          {
-            initialValue: 'month',
-          },
-        ],
-        // 小时：分钟
-        hourMinute: [
-          'hourMinute',
-          {
-            // initialValue: [1],
-            rules: [
-              { required: true, message: this.$t('cloudenv.text_446') },
-            ],
-          },
-        ],
-        // 有效时间
-        startEndTime: [
-          'startEndTime',
-          {
-            // initialValue: [1],
-            rules: [
-              { required: false, message: this.$t('cloudenv.text_447') },
-            ],
-          },
-        ],
-        execTime: [
-          'timer.execTime',
-          {
-            initialValue: this.$moment().add(1, 'h'),
-            rules: [
-              { required: true, message: this.$t('cloudenv.text_437') },
             ],
           },
         ],
@@ -230,42 +109,7 @@ export default {
         operation: 'sync',
         generate_name: values.name,
       }
-      if (values.cycleTimer.cycle_type === 'one') {
-        params.scheduled_type = 'timing'
-        params.timer = values.timer
-      } else {
-        params.scheduled_type = 'cycle'
-        params.cycle_timer = { ...values.cycleTimer }
-      }
-      // 触发时间
-      if (values.hourMinute) {
-        // 转换为utc hour
-        const time = this.$moment(values.hourMinute)
-        params.cycle_timer.hour = time.subtract(time.utcOffset(), 'minutes').hour()
-        params.cycle_timer.minute = values.hourMinute.minutes()
-      }
-      // 有效时间
-      if (values.startEndTime && values.startEndTime.length > 0) {
-        params.cycle_timer.startTime = values.startEndTime[0].set({
-          hour: 0,
-          minute: 0,
-          second: 0,
-        })
-        params.cycle_timer.endTime = values.startEndTime[1].set({
-          hour: 23,
-          minute: 59,
-          second: 59,
-        })
-      } else {
-        // 未设置有效时间时，有效时间为 今天-100年
-        if (!params.cycle_timer) params.cycle_timer = {}
-        params.cycle_timer.startTime = this.$moment()
-        params.cycle_timer.endTime = this.$moment().add('year', 100)
-      }
-      if (params.cycle_timer && params.cycle_timer.cycle_type === 'cycle') {
-        params.cycle_timer.cycle_type = params.cycle_timer.cycle_type2
-        delete params.cycle_timer.cycle_type2
-      }
+      this.$refs.taskTime.transformParams(params, values)
       return params
     },
     async handleConfirm () {
