+++ containers/Cloudenv/views/scheduledtask/create/index.vue
@@ -15,59 +15,7 @@
         <a-form-item :label="$t('common.description')">
           <a-textarea :auto-size="{ minRows: 1, maxRows: 3 }" v-decorator="decorators.description" :placeholder="$t('common_367')" />
         </a-form-item>
-        <a-form-item :label="$t('cloudenv.text_433')">
-          <a-radio-group v-decorator="decorators.cycle_type">
-            <a-radio-button v-for="(v, k) in $t('cloudenvScheduledtaskGroupCycleType')" :key="k" :value="k">{{v}}</a-radio-button>
-          </a-radio-group>
-        </a-form-item>
-        <!-- 周期策略 -->
-        <a-form-item class="d-flex" :label="$t('cloudenv.frequency')" v-if="form.fc.getFieldValue('cycleTimer.cycle_type') === 'cycle'">
-          <div class="d-flex align-items-center">
-            {{$t('cloudenv.each')}}
-            <a-input-number class="ml-2" :min="1" v-decorator="decorators.cycle_num" />
-            <a-select class="mr-2" v-decorator="decorators.cycle_type2" style="width: 100px">
-              <a-select-option v-for="(item,index) in cycleTypeOpts" :key="index" :value="item.key">{{item.label}}</a-select-option>
-            </a-select>
-            {{$t('cloudenv.cycle_once')}}
-          </div>
-        </a-form-item>
-        <!-- 周 -->
-        <a-form-item :label="$t('cloudenv.text_434')" v-if="form.fc.getFieldValue('cycleTimer.cycle_type') === 'week'">
-          <a-select v-decorator="decorators.weekDays" mode="multiple">
-            <a-select-option v-for="(v, k) in $t('flexGroupSubCycleTypeWeek')" :key="k" :value="parseInt(k)">{{v}}</a-select-option>
-          </a-select>
-        </a-form-item>
-        <!-- 月 -->
-        <a-form-item :label="$t('cloudenv.text_435')" v-if="form.fc.getFieldValue('cycleTimer.cycle_type') === 'month'">
-          <a-select v-decorator="decorators.monthDays" mode="multiple">
-            <a-select-option v-for="i in 31" :key="i" :value="parseInt(i)">{{$t('cloudenv.text_436', [i])}}</a-select-option>
-          </a-select>
-        </a-form-item>
-        <!-- 单次策略 -->
-        <template v-if="form.fc.getFieldValue('cycleTimer.cycle_type') === 'one'">
-          <a-form-item :label="$t('cloudenv.text_437')">
-            <a-date-picker
-              :showTime="{
-                format: 'HH:mm',
-              }"
-              :disabledDate="disabledDate"
-              :disabledTime="disabledDateTime"
-              v-decorator="decorators.execTime"
-              format="YYYY-MM-DD HH:mm" />
-          </a-form-item>
-        </template>
-        <!-- 非单次策略 -->
-        <template v-if="form.fc.getFieldValue('cycleTimer.cycle_type') !== 'one'">
-          <a-form-item :label="$t('cloudenv.text_437')" v-if="form.fc.getFieldValue('cycleTimer.cycle_type') !== 'cycle'">
-            <a-time-picker v-decorator="decorators.hourMinute" format="HH:mm" />
-          </a-form-item>
-          <a-form-item :label="$t('cloudenv.text_438')">
-            <a-range-picker
-              v-decorator="decorators.startEndTime"
-              :disabledDate="disabledDate"
-              format="YYYY-MM-DD" />
-          </a-form-item>
-        </template>
+        <scheduled-task-time ref="taskTime" :form="form" />
         <a-form-item :label="$t('cloudenv.text_384')">
           <a-select v-decorator="decorators.resourceType">
             <a-select-option v-for="(v, k) in resourceTypeOpts" :key="k" :value="k">{{v}}</a-select-option>
@@ -106,6 +54,7 @@
 
 <script>
 import { mapGetters } from 'vuex'
+import ScheduledTaskTime from '@Cloudenv/sections/ScheduledTaskTime'
 import Tag from '@/sections/Tag'
 import DomainProject from '@/sections/DomainProject'
 import ListSelect from '@/sections/ListSelect'
@@ -120,6 +69,7 @@ export default {
     Tag,
     ListSelect,
     NameRepeated,
+    ScheduledTaskTime,
   },
   mixins: [ResourcePropsMixin],
   data () {
@@ -291,12 +241,6 @@ export default {
           xxl: { span: 2 },
         },
       },
-      cycleTypeOpts: [
-        { key: 'hour', label: this.$t('cloudenv.cycle_type.hour') },
-        { key: 'day', label: this.$t('cloudenv.cycle_type.day') },
-        { key: 'week', label: this.$t('cloudenv.cycle_type.week') },
-        { key: 'month', label: this.$t('cloudenv.cycle_type.month') },
-      ],
     }
   },
   computed: {
@@ -316,31 +260,8 @@ export default {
       }
       return result
     },
-    disabledDateTime (_ = this.$moment(), type) {
-      let disabledHours = []
-      let disabledMinutes = []
-      let disabledSeconds = []
-      const dayDiff = _.diff(this.$moment(), 'days', true)
-      const hourDiff = _.diff(this.$moment(), 'hours', true)
-      const minutesDiff = _.diff(this.$moment(), 'minutes', true)
-      // 当天
-      if (dayDiff > -1 && dayDiff < 0) {
-        disabledHours = this.range(0, this.$moment().hour())
-        if (hourDiff > -1 && hourDiff < 0) {
-          disabledMinutes = this.range(0, this.$moment().minute())
-          if (minutesDiff > -1 && minutesDiff < 0) {
-            disabledSeconds = this.range(0, this.$moment().second())
-          }
-        }
-      }
-      return {
-        disabledHours: () => disabledHours,
-        disabledMinutes: () => disabledMinutes,
-        disabledSeconds: () => disabledSeconds,
-      }
-    },
     generateData (values) {
-      const { domain, project, name, description, action, labelType, servers, cycleTimer, resourceType } = values
+      const { domain, project, name, description, action, labelType, servers, resourceType } = values
       const params = {
         generate_name: name,
         description,
@@ -350,42 +271,7 @@ export default {
         operation: action,
         label_type: labelType,
       }
-      if (cycleTimer.cycle_type === 'one') {
-        params.scheduled_type = 'timing'
-        params.timer = values.timer
-      } else {
-        params.scheduled_type = 'cycle'
-        params.cycle_timer = { ...cycleTimer }
-      }
-      // 触发时间
-      if (values.hourMinute) {
-        // 转换为utc hour
-        const time = this.$moment(values.hourMinute)
-        params.cycle_timer.hour = time.subtract(time.utcOffset(), 'minutes').hour()
-        params.cycle_timer.minute = values.hourMinute.minutes()
-      }
-      // 有效时间
-      if (values.startEndTime && values.startEndTime.length > 0) {
-        params.cycle_timer.startTime = values.startEndTime[0].set({
-          hour: 0,
-          minute: 0,
-          second: 0,
-        })
-        params.cycle_timer.endTime = values.startEndTime[1].set({
-          hour: 23,
-          minute: 59,
-          second: 59,
-        })
-      } else {
-        // 未设置有效时间时，有效时间为 今天-100年
-        if (!params.cycle_timer) params.cycle_timer = {}
-        params.cycle_timer.startTime = this.$moment()
-        params.cycle_timer.endTime = this.$moment().add('year', 100)
-      }
-      if (params.cycle_timer && params.cycle_timer.cycle_type === 'cycle') {
-        params.cycle_timer.cycle_type = params.cycle_timer.cycle_type2
-        delete params.cycle_timer.cycle_type2
-      }
+      this.$refs.taskTime.transformParams(params, values)
       if (labelType === 'id') {
         params.labels = servers
       } else if (labelType === 'tag') {
